probability: assume frequencies to be independent
For dependent probabilities that sum to 1, we get the total probability
by calculating:

n! * \prod(q_i^{k_i} / k_i!)

If we want to compare two of these probabilities, we can leave out
anything that does not depend on q. With $p_i = k_i/n$, this results in:

\prod(q_i^{p_i})

This is what we have done so far. However, our probabilities do not sum to
1. If we only look at 1-grams, we can add an "everything else" factor:

(1 - \sum q_i)^{1 - sum p_i}

While 1-grams, 2-grams and 3-grams are not totally independent, they are
also not dependent in the sense that one implies not-the-other. So we
could caluclate the probability of each group individually and multiply
the results.

In practice, just treating everything as independent seems to work just
as well and is much simpler.